+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## tonic [0.4.1] - 2023-12-05
+
+### Added
+
+- Add new API in `tonic::transport::Server`.
+
 ## aws-sdk-s3 [0.4.0] - 2023-11-24
 
 ### Changed